Refactor the protocol support a bit
This change introduces GitProtocol, which contains the three Callbacks
and a function to push a packfile directly into a git.Repository. This
is done to increase composability and avoid having to create an HTTP
request if you already have a packfile handy (e.g. to push the contents
of a .zip file directly as a branch as if it were a commit).