Make the git browser also respect the protocol's ReferenceDiscoveryCallback
This change prevents the users from being able to discover hidden
references through the git browser. This is done by limiting the /+refs/
handler by only showing refs that the caller has permission to view, as
well as limiting the other handlers by checking whether the commit of
the requested resource is reachable by any of the refs that are
viewable.