Use git2go's ReachableFromAny() (#9)
This change replaces the MergeBase() call with a fancy new
ReachableFromAny(). Under the covers, ReachableFromAny() calls
MergeBaseMany(), but if there's a `commit-graph` file, it will use that
to optimize the search to prune it when it knows for sure that the
commit cannot be reachable thanks to the generation information.