Allow `/+archive/` to serve a tarball from a tree-id (#24)
Previously we only allowed commit IDs. But the grader does really need
to be able to access an arbitrary tree (in reality, the tree from a
particular branch that's a direct descendant of a commit, but that's too
complicated to express in a simple request). Furthermore, when operating
on behalf of the runners, it does not have access to the commit anymore!

This change now allows tree-ids to be specified instead of a commit.
This is also true now of `/+show/`, where we had a small vulnerability:
folks could sidestep the commit check by adding the `^{tree}` suffix to
the rev!